+++ src/tree/iterators.rs
@@ -38,7 +38,6 @@ pub struct LeafBit {
 }
 
 /// An iterator over a K2Tree's stems which produces instances of StemBit.
-#[derive(Debug)]
 pub struct Stems<'a> {
   tree: &'a K2Tree,
   pos: usize,
@@ -92,7 +91,6 @@ impl<'a> Stems<'a> {
 }
 
 /// A consuming iterator over a K2Tree's stems which produces instances of StemBit.
-#[derive(Debug)]
 pub struct IntoStems {
   tree: K2Tree,
   pos: usize,
@@ -146,7 +144,6 @@ impl IntoStems {
 }
 
 /// An iterator over a K2Tree's leaves which produces instances of LeafBit.
-#[derive(Debug)]
 pub struct Leaves<'a> {
   tree: &'a K2Tree,
   pos: usize,
@@ -180,7 +177,6 @@ impl<'a> Leaves<'a> {
 }
 
 /// A consuming iterator over a K2Tree's leaves which produces instances of LeafBit.
-#[derive(Debug)]
 pub struct IntoLeaves {
   tree: K2Tree,
   pos: usize,
@@ -215,7 +211,6 @@ impl IntoLeaves {
 }
 
 /// An iterator over a K2Tree's stems which produces the raw boolean-values of each bit.
-#[derive(Debug)]
 pub struct StemsRaw<'a> {
   stems: &'a BitVec,
   pos: usize,
@@ -242,7 +237,6 @@ impl<'a> StemsRaw<'a> {
 }
 
 /// An iterator over a K2Tree's leaves which prduces the raw boolean-values of each bit.
-#[derive(Debug)]
 pub struct LeavesRaw<'a> {
   leaves: &'a BitVec,
   pos: usize,
@@ -267,3 +261,12 @@ impl<'a> LeavesRaw<'a> {
     }
   }
 }
+
+// #[cfg(test)]
+// mod tests {
+//   use super::*;
+//   #[test]
+//   fn IntoStems() {
+//     let tree = K2Tree::
+//   }
+// }
