+++ src/tree/mod.rs
@@ -14,10 +14,12 @@ pub use iterators::{
   LeavesRaw,
 };
 
-/* Common */
-use bitvec::vec::BitVec;
+/* Private Yet Common to Everything Module */
+use bitvec::{prelude::{bitvec, bitbox, BitVec}};
+use crate::matrix::BitMatrix;
 
 impl K2Tree {
+  // Internal: Never call with layer >= self.max_slayers
   fn layer_len(&self, l: usize) -> usize {
     if l == self.max_slayers-1 {
       return self.stems.len() - self.layer_start(l)
@@ -336,5 +338,133 @@ mod range_tests {
     let subs = SubRanges::from_range(original, 3, 2)?;
     for i in 0..6 { assert_eq!(expected_subs[i], subs[i]); }
     Ok(())
+  }
+}
+
+/* Private funcs used in testing */
+#[cfg(test)]
+impl K2Tree {
+  fn test_tree(k: usize) -> Self {
+    match k {
+      2 => K2Tree {
+        stem_k: 2,
+        leaf_k: 2,
+        max_slayers: 2,
+        stems:  bitvec![0,1,1,1, 1,1,0,1, 1,0,0,0, 1,0,0,0],
+        leaves: bitvec![0,1,1,0, 0,1,0,1, 1,1,0,0, 1,0,0,0, 0,1,1,0],
+      },
+      3 => K2Tree {
+        stem_k: 3,
+        leaf_k: 3,
+        max_slayers: 2,
+        stems:  bitvec![
+          0,1,0,1,1,0,1,1,0, 1,1,0,0,0,0,0,0,0, 1,0,0,0,0,0,0,0,0,
+          1,0,0,0,0,0,0,0,0, 1,0,0,0,0,0,0,0,0, 1,0,0,0,0,0,0,0,0
+        ],
+        leaves: bitvec![
+          0,1,0,1,0,0,0,0,1, 1,0,0,1,0,0,1,0,0, 1,0,0,0,0,0,0,0,0,
+          0,1,0,1,0,0,0,0,0, 1,0,0,0,0,0,0,0,0, 0,1,0,1,0,0,0,0,0,
+        ]
+      },
+      4 => K2Tree {
+        stem_k: 4,
+        leaf_k: 4,
+        max_slayers: 2,
+        stems: bitvec![
+          1,0,0,1,0,0,0,1,1,0,0,0,1,1,0,1, 1,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,
+          0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0, 0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,
+          0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0, 0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,
+          1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0, 0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,
+        ],
+        leaves: bitvec![
+          1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0, 0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,
+          0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0, 0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,
+          0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0, 0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,
+          0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,
+        ],
+      },
+      _ => K2Tree::with_k(2, 2).unwrap(),
+    }
+  }
+  fn test_matrix(k: usize) -> BitMatrix {
+    match k {
+      2 => {
+        let bits = bitvec![
+          0,0,0,0, 0,1,0,1,
+          0,0,0,0, 1,0,0,1,
+          0,0,0,0, 0,0,1,1,
+          0,0,0,0, 0,0,0,0,
+
+          1,0,0,0, 0,1,0,0,
+          0,0,0,0, 1,0,0,0,
+          0,0,0,0, 0,0,0,0,
+          0,0,0,0, 0,0,0,0
+        ];
+        BitMatrix::from_bits(8, 8, bits)
+      },
+      3 => {
+        let bits = bitvec![
+          0,0,0, 0,0,0, 0,0,0,  0,1,0, 1,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  1,0,0, 1,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,1, 1,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+
+          1,0,0, 0,0,0, 0,0,0,  0,1,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  1,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          
+          1,0,0, 0,0,0, 0,0,0,  0,1,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  1,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+          0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,  0,0,0, 0,0,0, 0,0,0,
+        ];
+        BitMatrix::from_bits(27, 27, bits)
+      },
+      _ => BitMatrix::new(),
+    }
+  }
+}
+
+#[cfg(test)]
+mod K2Tree_common_util {
+  use super::*;
+  #[test]
+  fn layer_len() {
+    let tree = K2Tree::test_tree(2);
+    assert_eq!(4, tree.layer_len(0));
+    assert_eq!(12, tree.layer_len(1));
+  }
+  #[test]
+  fn get_coords() {
+    let expected_coords = [
+      [4, 0], [5, 0], [4, 1], [5, 1], // Leaf 1
+      [6, 0], [7, 0], [6, 1], [7, 1], // Leaf 2
+      [6, 2], [7, 2], [6, 3], [7, 3], // Leaf 3
+      [0, 4], [1, 4], [0, 5], [1, 5], // Leaf 4
+      [4, 4], [5, 4], [4, 5], [5, 5], // Leaf 5
+    ];
+    let tree = K2Tree::test_tree(2);
+    for (leaf_offset, _) in tree.leaves.iter().enumerate() {
+      assert_eq!(
+        expected_coords[leaf_offset],
+        tree.get_coords(leaf_offset)
+      );
+    }
   }
 }
